net: ipconfig: Support using "delayed" DHCP replies
commit 2647cffb2bc6fbed163d377390eb7ca552c7c1cb upstream.

The dhcp code only waits 1s between sending DHCP requests on different
devices and only accepts an answer for the device that sent out the last
request. Only the timeout at the end of a loop is increased iteratively
which favours only the last device. This makes it impossible to work
with a dhcp server that takes little more than 1s connected to a device
that is not the last one.

Instead of also increasing the inter-device timeout, teach the code to
handle delayed replies.

To accomplish that, make *ic_dev track the current ic_device instead of
the current net_device and adapt all users accordingly. The relevant
change then is to reset d to ic_dev on a reply to assert that the
followup request goes through the right device.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Signed-off-by: David S. Miller <davem@davemloft.net>
[Patryk: Backported to 4.4: replaced DBG function with pr_debug
squashed upstream commits:
e068853409aa17208e94f4ca628005cc6f51e043
9c706a49d660653625d206f6972541c1f60ea2b0
d2d371ae5dd6af9a6a3d7f50b753627c42868409
1ae292a2457cd692828da2be87cb967260993ad0]
Signed-off-by: Patryk Mungai <patryk.mungai-ndungu.kx@renesas.com>
Signed-off-by: Pavel Machek <pavel@denx.de>