usb: host: xhci-plat: Add r8a7744 support
This patch adds xhci r8a7744 driver support.

NOTE: Just a background wrt this patch as this isn't a backport from mainline
kernel,  in mainline kernel for RZ/G1 the compatible string defaults to
"renesas,rcar-gen2-xhci"  which isn't present in CIP, and the way
xhci-plat.c in cip handles compatibility is different compared to mainline as
a result this patch is added.
This is patch is similar to  "usb: host: xhci-plat: Add r8a7743
support" (d09ccbb733864223402b541bdb9897cfe811881c) present in cip
kernel.

Signed-off-by: Lad Prabhakar <prabhakar.mahadev-lad.rj@bp.renesas.com>
Signed-off-by: Nobuhiro Iwamatsu <nobuhiro1.iwamatsu@toshiba.co.jp>